Allow gem name to be specified in bower.json.
With this change, the rake component:convert task can now be used to generate
gems using private bower packages hosted anywhere, even in a local .git
repository. (Normally rails-assets only works with publicly-registered bower
packages or public GitHub repositories.)

For this to work, a "gem" property has to be present in the bower.json. It
specifies the desired name of the gem to be generated.